Rework sensor
* Moved EmailData class to __init__.py
